MEOW MEOW MEOW auuuuuuhhhihihihhohuhuho

+++ MarketBasket/MarketBasket/Triple.cs
@@ -9,13 +9,13 @@ namespace MarketBasket
     public class Triple
     {
         public int[] Ids { get; set; }
-        public int?[] Scores { get; set; }
+        public double?[] Scores { get; set; }
         private int[] Counts { get; set; }
 
         public Triple(int item1, int item2, int item3)
         {
             Ids = new int[3];
-            Scores = new int?[7];
+            Scores = new double?[7];
             Counts = new int[7];
             Ids[0] = item1;
             Ids[1] = item2;
@@ -25,7 +25,7 @@ namespace MarketBasket
         public void AddSentiment(int score, int dayValue)
         {
             if (Scores[dayValue] == null) Scores[dayValue] = 0;
-            Scores[dayValue] = (Scores[dayValue] * Counts[dayValue] + score) / (Counts[dayValue] + 1);
+            Scores[dayValue] = (Scores[dayValue] * Counts[dayValue] + score) / (Counts[dayValue] + 1.0);
             Counts[dayValue]++;
         }
     }
